CSCI3130-15 [US:7] Set up unit tests for logging in
Created a unit test class for our LoginView. For the moment, we cannot
implement unit tests on this class because it is a view object, whose
inner classes and functions are protected and thus cannot be accessed
for testing. The LoginViewTest class is meant as an empty container
to host future test functions, should we need them.

Additionally, pulled the login function out from its Button. It is
now a generic function that we can utilize and call from anywhere we
see fit.

jira: [CSCI3130-15]