CSCI3130-14 - [US:7] Link Log-In Fields with data from sign-up
Lot of stuff in this. Modified the LoginView to center the text
fields and added in a keypress listener (shortcut listener) for when
the ENTER key is pressed. This is in line with

The core change in this is the ability for users to properly sign
in with a new account that they have created, i.e. ensuring. that the
persisted database can be added to and accessed through the UI when
a user is trying to log in. To this end, we've refactored the
RegisteredUserDatabase to include a fetchUser function, through which
we can find a user and do a validation on their username and password.

Future: refactor so that our back-end classes and functions (like
RegisteredUser and ESPECIALLY RegisteredUserDatabase) are held in a
back end and thus more modular.

Authors: Jonathan McDevitt, Nick Broderick